requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
deepdiff==8.0.1
fulcrum==1.12.0
idna==3.10
orderly-set==5.2.2
PyMuPDF==1.24.14
python-dotenv==1.0.1
requests==2.32.3
tabulate==0.9.0
tqdm==4.65.0
urllib3==2.2.3
